Use subclasses to expose types of servers
These changes break the existing ServerApp into 3 subclasses of
BaseServerApp - ServerApp (for serving notebooks), KernelServerApp
(for serving kernels) and ContentsServerApp (for serving content).

This approach allows for more specific behaviors in the various
subclasses, but, I suspect, breaks the ExtensionApp modeling, since
that appears to assume extensions are applied to ServerApp.  As a
result, more investigation is required if we wanted KernelServerApp
or ContentsServerApp to have the ability to be extended, rather than
only subclassed.

All tests currently pass against ServerApp, and KernelServerApp appears
to startup fine and only serve the application endpoints.