Use server-mode option to expose types of servers
These changes use a new option `--server-mode` to indicate which mode
the server will run in.  The available modes are 'notebooks' (default),
'kernels' and 'contents'.   The current behavior is exhibited with
'notebooks'.  However, when either 'kernels' or 'contents' are indicated,
only the services (REST API) relative to those modes will be available.

Because all modes are implemented in the same class, this approache (vs.
subclassing) shouldn't side-affect the ability for server extensions to
be written against the various modes.