Use subclasses composed of service-related mixins
These changes, like those in the 'service-composition-w-classes' branch
break ServerApp into a BaseServerApp class that ServerApp derives from.
However, it introduces Mixin classes for each of the services and also
which ServerApp derives from.  This approach allows developers to build
an application by composing service-related Mixins.

Because some mixins depend on others, a dependency validation step has
been introduced to ensure appropriate dependencies are in place.

The current tests pass, but I suspect there are issues with aliases.
These same issues also exist in the 'w-classes' branch (and I'm not
sure about the 'w-options' branch - but I doubt it given that approach).